requirements.txt
openai
python-dotenv
python-docx
python-pptx
PyMuPDF
aiohttp
aiofiles
tqdm
tiktoken